heatwave3: Detect Heatwaves and Cold-Spells within cubed data

This package is an extension to heatwaveR that allows the user to detect extreme events a la Hobday definition directly on cubed data files (e.g. NetCDF format).

See also

Author

Maintainer: Robert W. Schlegel robwschlegel@gmail.com (ORCID) [contributor]

Authors:

Guilherme Pinto [contributor]

Lily Genevier [contributor]

Jacob Cohen [contributor]

Panini Dasgupta [contributor]

Evan Meeker [contributor]

Elise Beaudin [contributor]

Albertus J. Smit (ORCID) [contributor]